+++ MaxMind.MinFraud/Exception/MinFraudException.cs
@@ -7,7 +7,6 @@ namespace MaxMind.MinFraud.Exception
     /// This class represents a non-specific error with data returned by
     /// the web service.
     /// </summary>
-    [Serializable]
     public class MinFraudException : System.Exception
     {
         /// <summary>
@@ -31,16 +30,6 @@ namespace MaxMind.MinFraud.Exception
         /// <param name="message">Exception message.</param>
         /// <param name="innerException">The underlying exception that caused this one.</param>
         public MinFraudException(string message, System.Exception innerException) : base(message, innerException)
-        {
-        }
-
-        /// <summary>
-        /// Constructor for deserialization.
-        /// </summary>
-        /// <param name="info">The SerializationInfo with data.</param>
-        /// <param name="context">The source for this deserialization.</param>
-        protected MinFraudException(SerializationInfo info, StreamingContext context)
-            : base(info, context)
         {
         }
     }